@@ -23,29 +23,30 @@
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
-    "php": "^7.3|^8.0",
+    "php": "^7.4|^8.0",
     "slim/slim": "^4.7",
     "willdurand/negotiation": "^3.0"
   },
   "require-dev": {
-    "dealerdirect/phpcodesniffer-composer-installer": "^0.7",
+    "dealerdirect/phpcodesniffer-composer-installer": "~0.7",
+    "grifart/phpstan-oneline": "~0.4",
     "filp/whoops": "^2.8",
-    "friendsofphp/php-cs-fixer": "^2.17",
-    "infection/infection": "^0.18|^0.21",
+    "infection/infection": "~0.24",
+    "juliangut/php-cs-fixer-config": "dev-master",
     "laminas/laminas-diactoros": "^2.2",
-    "overtrue/phplint": "^2.3",
-    "phpmd/phpmd": "^2.9",
+    "overtrue/phplint": "^3.0|^4.0",
     "phpcompatibility/php-compatibility": "^9.3",
-    "phpstan/extension-installer": "^1.1.0",
-    "phpstan/phpstan": "^0.12",
-    "phpstan/phpstan-deprecation-rules": "^0.12",
-    "phpstan/phpstan-strict-rules": "^0.12",
+    "phpmd/phpmd": "^2.10",
+    "phpstan/extension-installer": "^1.1",
+    "phpstan/phpstan": "^1.0",
+    "phpstan/phpstan-deprecation-rules": "^1.0",
+    "phpstan/phpstan-strict-rules": "^1.0",
     "phpunit/phpunit": "^9.5",
-    "povils/phpmnd": "^2.4",
+    "povils/phpmnd": "^2.5",
     "roave/security-advisories": "dev-master",
-    "sebastian/phpcpd": "^6.0.3",
-    "squizlabs/php_codesniffer": "^3.5",
-    "thecodingmachine/phpstan-strict-rules": "^0.12"
+    "sebastian/phpcpd": "^6.0",
+    "squizlabs/php_codesniffer": "^3.6",
+    "thecodingmachine/phpstan-strict-rules": "^1.0"
   },
   "suggest": {
     "filp/whoops": "Enhance development error reporting and production stack-traces",
@@ -64,6 +65,11 @@
   },
   "config": {
     "preferred-install": "dist",
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true,
+      "infection/extension-installer": true,
+      "phpstan/extension-installer": true
+    }
   }
 }
