@@ -15,10 +15,8 @@ namespace Jgut\Slim\Exception\Renderer;
 
 use Slim\Exception\HttpException;
 use Slim\Interfaces\ErrorRendererInterface;
+use Throwable;
 
-/**
- * Abstract exception renderer.
- */
 abstract class AbstractRenderer implements ErrorRendererInterface
 {
     /**
@@ -31,12 +29,6 @@ abstract class AbstractRenderer implements ErrorRendererInterface
      */
     protected $defaultDescription;
 
-    /**
-     * Exception renderer constructor.
-     *
-     * @param string $defaultTitle
-     * @param string $defaultDescription
-     */
     public function __construct(
         string $defaultTitle = 'Slim Application error',
         string $defaultDescription = 'A website error has occurred. Sorry for the temporary inconvenience.'
@@ -47,24 +39,16 @@ abstract class AbstractRenderer implements ErrorRendererInterface
 
     /**
      * Get exception title.
-     *
-     * @param \Throwable $exception
-     *
-     * @return string
      */
-    protected function getErrorTitle(\Throwable $exception): string
+    protected function getErrorTitle(Throwable $exception): string
     {
         return $exception instanceof HttpException ? $exception->getTitle() : $this->defaultTitle;
     }
 
     /**
      * Get exception description.
-     *
-     * @param \Throwable $exception
-     *
-     * @return string
      */
-    protected function getErrorDescription(\Throwable $exception): string
+    protected function getErrorDescription(Throwable $exception): string
     {
         if (!$exception instanceof HttpException) {
             return $this->defaultDescription;
