@@ -13,72 +13,66 @@ declare(strict_types=1);
 
 namespace Jgut\Slim\Exception\Renderer;
 
-/**
- * HTML exception renderer.
- */
+use Throwable;
+
 class HtmlRenderer extends AbstractRenderer
 {
     /**
-     * {@inheritdoc}
+     * @inheritDoc
      */
-    public function __invoke(\Throwable $exception, bool $displayErrorDetails): string
+    public function __invoke(Throwable $exception, bool $displayErrorDetails): string
     {
         $title = $this->getErrorTitle($exception);
         if ($displayErrorDetails) {
             $content = <<<CONTENT
-<p>The application could not run because of the following error:</p>
-<h2>Details</h2>
-{$this->formatException($exception)}
-CONTENT;
+            <p>The application could not run because of the following error:</p>
+            <h2>Details</h2>
+            {$this->formatException($exception)}
+            CONTENT;
         } else {
             $content = '<p>' . $this->getErrorDescription($exception) . '</p>';
         }
 
         return <<<OUTPUT
-<html lang="en">' .
-    <head>
-        <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-        <title>$title</title>
-        <style>
-            body{margin:0;padding:30px;font:12px/1.5 Helvetica,Arial,Verdana,sans-serif}
-            h1{margin:0;font-size:48px;font-weight:normal;line-height:48px}
-            strong{display:inline-block;width:65px}
-        </style>
-    </head>
-    <body>
-        <h1>$title</h1>
-        <div>$content</div>
-        <a href="#" onClick="window.history.go(-1)">Go Back</a>
-    </body>
-</html>
-OUTPUT;
+        <html lang="en">' .
+            <head>
+                <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+                <title>{$title}</title>
+                <style>
+                    body{margin:0;padding:30px;font:12px/1.5 Helvetica,Arial,Verdana,sans-serif}
+                    h1{margin:0;font-size:48px;font-weight:normal;line-height:48px}
+                    strong{display:inline-block;width:65px}
+                </style>
+            </head>
+            <body>
+                <h1>{$title}</h1>
+                <div>{$content}</div>
+                <a href="#" onClick="window.history.go(-1)">Go Back</a>
+            </body>
+        </html>
+        OUTPUT;
     }
 
-    /**
-     * @param \Throwable $exception
-     *
-     * @return string
-     */
-    private function formatException(\Throwable $exception): string
+    private function formatException(Throwable $exception): string
     {
         $outputString = <<<'OUTPUT'
-<div><strong>Type:</strong> %s</div>
-<div><strong>Code:</strong> %s</div>
-<div><strong>Message:</strong> %s</div>
-<div><strong>File:</strong> %s</div>
-<div><strong>Line:</strong> %s</div>
-<h2>Trace</h2>
-<pre>%s</pre>
-OUTPUT;
+        <div><strong>Type:</strong> %s</div>
+        <div><strong>Code:</strong> %s</div>
+        <div><strong>Message:</strong> %s</div>
+        <div><strong>File:</strong> %s</div>
+        <div><strong>Line:</strong> %s</div>
+        <h2>Trace</h2>
+        <pre>%s</pre>
+        OUTPUT;
 
-        return \sprintf(
+        return sprintf(
             $outputString,
             \get_class($exception),
             $exception->getCode(),
-            \htmlentities($exception->getMessage()),
+            htmlentities($exception->getMessage()),
             $exception->getFile(),
             $exception->getLine(),
-            \htmlentities($exception->getTraceAsString())
+            htmlentities($exception->getTraceAsString()),
         );
     }
 }
