@@ -16,10 +16,10 @@ namespace Jgut\Slim\Exception\Whoops\Renderer;
 use Jgut\Slim\Exception\Whoops\Inspector;
 use Whoops\Handler\Handler;
 use Whoops\Handler\XmlResponseHandler;
+use DOMDocument;
+use DOMElement;
+use SimpleXMLElement;
 
-/**
- * Whoops custom XML exception renderer.
- */
 class XmlRenderer extends XmlResponseHandler
 {
     use RendererTrait;
@@ -29,11 +29,6 @@ class XmlRenderer extends XmlResponseHandler
      */
     protected $prettify = true;
 
-    /**
-     * XmlHandler constructor.
-     *
-     * @param string $defaultTitle
-     */
     public function __construct(string $defaultTitle = 'Slim Application error')
     {
         $this->defaultTitle = $defaultTitle;
@@ -41,16 +36,13 @@ class XmlRenderer extends XmlResponseHandler
         $this->addTraceToOutput(true);
     }
 
-    /**
-     * @param bool $prettify
-     */
     public function setPrettify(bool $prettify): void
     {
         $this->prettify = $prettify;
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritDoc
      */
     public function handle()
     {
@@ -72,20 +64,18 @@ class XmlRenderer extends XmlResponseHandler
     /**
      * Get formatted XML exception.
      *
-     * @param mixed[] $data
-     *
-     * @return string
+     * @param array<mixed> $data
      */
     protected function getFormattedXml(array $data): string
     {
         $xmlTemplate = '<?xml version="1.0" encoding="UTF-8" standalone="yes"?><error />';
 
-        /** @var \SimpleXMLElement $root */
-        $root = \simplexml_load_string($xmlTemplate);
+        /** @var SimpleXMLElement $root */
+        $root = simplexml_load_string($xmlTemplate);
         $this->addDataNodes($root, $data, 'exception');
 
-        /** @var \SimpleXMLElement $rootDocument */
-        $rootDocument = \dom_import_simplexml($root);
+        /** @var SimpleXMLElement $rootDocument */
+        $rootDocument = dom_import_simplexml($root);
         $dom = $rootDocument->ownerDocument;
         $dom->formatOutput = $this->prettify;
 
@@ -97,36 +87,32 @@ class XmlRenderer extends XmlResponseHandler
     /**
      * Transform data to XML nodes.
      *
-     * @param \SimpleXMLElement $node
-     * @param mixed[]           $data
-     * @param string            $nodeKey
-     *
-     * @return \SimpleXMLElement
+     * @param array<mixed> $data
      */
-    protected function addDataNodes(\SimpleXMLElement $node, array $data, string $nodeKey): \SimpleXMLElement
+    protected function addDataNodes(SimpleXMLElement $node, array $data, string $nodeKey): SimpleXMLElement
     {
         foreach ($data as $key => $value) {
-            if (\is_numeric($key)) {
+            if (is_numeric($key)) {
                 $key = $nodeKey . '_' . $key;
             }
             /** @var string $key */
-            $key = \preg_replace('/[^a-z0-9\-_.:]/i', '_', $key);
+            $key = preg_replace('/[^a-z0-9\-_.:]/i', '_', $key);
 
             if (\is_array($value)) {
                 $this->addDataNodes($node->addChild($key), $value, $key);
             } else {
                 if (\is_object($value)) {
                     $value = \get_class($value);
-                } elseif (!\is_scalar($value)) {
+                } elseif (!is_scalar($value)) {
                     $value = \gettype($value);
                 }
 
-                $value = \str_replace('&', '&amp;', \print_r($value, true));
+                $value = str_replace('&', '&amp;', print_r($value, true));
 
                 if ($key === 'message') {
-                    /** @var \DOMElement $child */
-                    $child = \dom_import_simplexml($node->addChild($key));
-                    /** @var \DOMDocument $document */
+                    /** @var DOMElement $child */
+                    $child = dom_import_simplexml($node->addChild($key));
+                    /** @var DOMDocument $document */
                     $document = $child->ownerDocument;
 
                     $child->appendChild($document->createCDATASection($value));
